Updated light behavior
The light is the opposite of the pin (pin is off -> light is on, and vice versa), so code was changed to reflect this. The light now is off when the program starts, turns on when is requested to turn on, and turns off when is requested to turn off.